inproc connections emulations
Note that you could not connect to native 'inproc' connections.
You could treat it is as follow: 'inproc' connections are internal
to ZMQ context, here you've got internal "ZMQ context" which belongs
to EventMachine, so that you could not connect to native ZMQ 'inproc'
connections cause they are in other ZMQ context.